+++ Sofa/Component/ODESolver/Backward/CMakeLists.txt
@@ -6,12 +6,14 @@ set(SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR "src/sofa/component/odesolver/back
 set(HEADER_FILES
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/config.h.in
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/init.h
+    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/BDF1.h
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/EulerImplicitSolver.h
-    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/StaticSolver.h
-    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/VariationalSymplecticSolver.h
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/NewmarkImplicitSolver.h
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/NewtonRaphsonSolver.h
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/NewtonStatus.h
+    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/Static.h
+    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/StaticSolver.h
+    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/VariationalSymplecticSolver.h
     
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/convergence/AbsoluteConvergenceMeasure.h
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/convergence/AbsoluteEstimateDifferenceMeasure.h
@@ -23,11 +25,13 @@ set(HEADER_FILES
 
 set(SOURCE_FILES
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/init.cpp
+    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/BDF1.cpp
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/EulerImplicitSolver.cpp
-    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/StaticSolver.cpp
-    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/VariationalSymplecticSolver.cpp
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/NewmarkImplicitSolver.cpp
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/NewtonRaphsonSolver.cpp
+    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/Static.cpp
+    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/StaticSolver.cpp
+    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/VariationalSymplecticSolver.cpp
 )
 
 sofa_find_package(Sofa.Simulation.Core REQUIRED)
