+++ Sofa/Component/ODESolver/Backward/src/sofa/component/odesolver/backward/Static.h
@@ -20,9 +20,36 @@
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
 #pragma once
-#include <sofa/component/odesolver/integration/config.h>
+#include <sofa/component/odesolver/backward/config.h>
+#include <sofa/core/behavior/BaseIntegrationMethod.h>
+
+#include "sofa/core/behavior/MultiVec.h"
+#include "sofa/simulation/VectorOperations.h"
 
 namespace sofa::component::odesolver::integration
 {
-    SOFA_COMPONENT_ODESOLVER_INTEGRATION_API void init();
+class SOFA_COMPONENT_ODESOLVER_BACKWARD_API Static : public sofa::core::behavior::BaseIntegrationMethod
+{
+public:
+    SOFA_CLASS(Static, sofa::core::behavior::BaseIntegrationMethod);
+
+    std::size_t stepSize() const override;
+    void initializeVectors(const core::ExecParams* params, core::ConstMultiVecCoordId x, core::ConstMultiVecDerivId v) override;
+    Factors getMatricesFactors(SReal dt) const override;
+    void computeRightHandSide(const core::ExecParams* params, core::behavior::RHSInput input,
+                              core::MultiVecDerivId force, core::MultiVecDerivId rightHandSide,
+                              SReal dt) override;
+    void updateStates(const core::ExecParams* params, SReal dt, core::MultiVecCoordId x,
+                      core::MultiVecDerivId v, core::MultiVecCoordId newX,
+                      core::MultiVecDerivId newV,
+                      core::MultiVecDerivId linearSystemSolution) override;
+    SReal computeResidual(const core::ExecParams* params,
+        SReal dt,
+        core::MultiVecDerivId force,
+        core::MultiVecDerivId oldVelocity,
+        core::MultiVecDerivId newVelocity) override;
+
+private:
+    core::MultiVecCoordId x_i; //x[i]
+};
 }
