--- Sofa/Component/ODESolver/Integration/CMakeLists.txt
@@ -1,37 +1,0 @@
-cmake_minimum_required(VERSION 3.22)
-project(Sofa.Component.ODESolver.Integration LANGUAGES CXX)
-
-set(SOFACOMPONENTODESOLVERINTEGRATION_SOURCE_DIR "src/sofa/component/odesolver/integration")
-
-set(HEADER_FILES
-    ${SOFACOMPONENTODESOLVERINTEGRATION_SOURCE_DIR}/config.h.in
-    ${SOFACOMPONENTODESOLVERINTEGRATION_SOURCE_DIR}/init.h
-
-    ${SOFACOMPONENTODESOLVERINTEGRATION_SOURCE_DIR}/BDF1.h
-    ${SOFACOMPONENTODESOLVERINTEGRATION_SOURCE_DIR}/Static.h
-)
-
-set(SOURCE_FILES
-    ${SOFACOMPONENTODESOLVERINTEGRATION_SOURCE_DIR}/init.cpp
-
-    ${SOFACOMPONENTODESOLVERINTEGRATION_SOURCE_DIR}/BDF1.cpp
-    ${SOFACOMPONENTODESOLVERINTEGRATION_SOURCE_DIR}/Static.cpp
-)
-
-sofa_find_package(Sofa.Simulation.Core REQUIRED)
-
-add_library(${PROJECT_NAME} SHARED ${HEADER_FILES} ${SOURCE_FILES})
-target_link_libraries(${PROJECT_NAME} PUBLIC Sofa.Simulation.Core)
-
-sofa_create_package_with_targets(
-    PACKAGE_NAME ${PROJECT_NAME}
-    PACKAGE_VERSION ${Sofa_VERSION}
-    TARGETS ${PROJECT_NAME} AUTO_SET_TARGET_PROPERTIES
-    INCLUDE_SOURCE_DIR "src"
-    INCLUDE_INSTALL_DIR "${PROJECT_NAME}"
-)
-
-#cmake_dependent_option(SOFA_COMPONENT_ODESOLVER_INTEGRATION_BUILD_TESTS "Compile the automatic tests" ON "SOFA_BUILD_TESTS OR NOT DEFINED SOFA_BUILD_TESTS" OFF)
-#if(SOFA_COMPONENT_ODESOLVER_INTEGRATION_BUILD_TESTS)
-#    add_subdirectory(tests)
-#endif()
