--- Sofa/Component/ODESolver/Integration/Sofa.Component.ODESolver.IntegrationConfig.cmake.in
@@ -1,12 +1,0 @@
-# CMake package configuration file for the @PROJECT_NAME@ module
-
-@PACKAGE_GUARD@
-@PACKAGE_INIT@
-
-find_package(Sofa.Simulation.Core QUIET REQUIRED)
-
-if(NOT TARGET @PROJECT_NAME@)
-    include("${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@Targets.cmake")
-endif()
-
-check_required_components(@PROJECT_NAME@)
