--- Sofa/Component/ODESolver/Integration/src/sofa/component/odesolver/integration/config.h.in
@@ -1,38 +1,0 @@
-/******************************************************************************
-*                 SOFA, Simulation Open-Framework Architecture                *
-*                    (c) 2006 INRIA, USTL, UJF, CNRS, MGH                     *
-*                                                                             *
-* This program is free software; you can redistribute it and/or modify it     *
-* under the terms of the GNU Lesser General Public License as published by    *
-* the Free Software Foundation; either version 2.1 of the License, or (at     *
-* your option) any later version.                                             *
-*                                                                             *
-* This program is distributed in the hope that it will be useful, but WITHOUT *
-* ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or       *
-* FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License *
-* for more details.                                                           *
-*                                                                             *
-* You should have received a copy of the GNU Lesser General Public License    *
-* along with this program. If not, see <http://www.gnu.org/licenses/>.        *
-*******************************************************************************
-* Authors: The SOFA Team and external contributors (see Authors.txt)          *
-*                                                                             *
-* Contact information: contact@sofa-framework.org                             *
-******************************************************************************/
-#pragma once
-
-#include <sofa/config.h>
-
-#ifdef SOFA_BUILD_SOFA_COMPONENT_ODESOLVER_INTEGRATION
-#  define SOFA_TARGET @PROJECT_NAME@
-#  define SOFA_COMPONENT_ODESOLVER_INTEGRATION_API SOFA_EXPORT_DYNAMIC_LIBRARY
-#else
-#  define SOFA_COMPONENT_ODESOLVER_INTEGRATION_API SOFA_IMPORT_DYNAMIC_LIBRARY
-#endif
-
-
-namespace sofa::component::odesolver::integration
-{
-	constexpr const char* MODULE_NAME = "@PROJECT_NAME@";
-	constexpr const char* MODULE_VERSION = "@PROJECT_VERSION@";
-} // namespace sofa::component::odesolver::integration
